+++ src/banner.rs
@@ -16,12 +16,12 @@ const HALVING_INTERVAL: u64 = 210_000;
 pub fn get_welcome_banner(query: &Query, omit_donation: bool) -> Result<String> {
     let rpc = query.rpc();
 
-    let net_info = rpc.get_network_info_()?;
+    let net_info = rpc.get_network_info()?;
     let chain_info = rpc.get_blockchain_info()?;
     let mempool_info = rpc.get_mempool_info()?;
     let net_totals = rpc.get_net_totals()?;
     let peers = rpc.get_peer_info()?;
-    let hash_rate_7d = rpc.get_network_hash_ps(1008)?;
+    let hash_rate_7d = rpc.get_network_hash_ps(Some(1008), None)?;
     let uptime = dur_from_secs(rpc.uptime()?);
     let tip = rpc.get_block_stats(&rpc.get_best_block_hash()?)?;